v0.7.9
* `restore_particle_ordering!` added, this restores optimal indexing of particles and can lead to simulation speed-ups due to improved cache usage
* Minor optimizations in octree merging
* Simplified elastic VHS collision code via a unified `collide_2particles_vhs!` function, minor speed-ups
* New keyword parameter `dw_tol` in collision routines that sets tolerance in weight difference under which particle
collisions are treated as equal weight collisions and no additional particle splitting is performed
* Minor improvements in readability and speed of electron-neutral collision routines
* `@inbounds` added to acceleration routine
* `ntc_equal_weight` added for slightly faster collisions in equal-weight simulations
* Documentation improvements
* Improved test coverage

+++ src/distributions_and_sampling.jl
@@ -20,9 +20,9 @@ Stores information about a uniform discrete 3-dimensional velocity grid with ext
 * `vz_grid`: `Vector` of grid nodes in z direction
 """
 struct UnitDVGrid
-    nx::Int16
-    ny::Int16
-    nz::Int16
+    nx::Int64
+    ny::Int64
+    nz::Int64
     dx::Float64
     dy::Float64
     dz::Float64
@@ -74,9 +74,9 @@ Stores the values of a function evaluated on 3-dimensional velocity grid.
 * `w`: the 3-dimensional array of values
 """
 mutable struct VDF
-    const nx::Int16
-    const ny::Int16
-    const nz::Int16
+    const nx::Int64
+    const ny::Int64
+    const nz::Int64
     w::Array{Float64,3}
 end
